% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_lib_path.R
\name{add_lib_path}
\alias{add_lib_path}
\title{Add library path}
\usage{
add_lib_path(libPath)
}
\arguments{
\item{libPath}{The path to add, as a character string.}
}
\value{
Resulting value of \code{.libPaths()}
}
\description{
\code{add_lib_path} adds a path to the front of \code{.libPaths()}, making the added
path be the first path checked for installed packages.
}
\details{
Does nothing if the passed library path is already the first path in
\code{.libPaths()}.
}
