% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_element_length_is_one.R
\name{assert_element_length_is_one}
\alias{assert_element_length_is_one}
\title{Assert that each element in the list is of length 1.}
\usage{
assert_element_length_is_one(x)
}
\arguments{
\item{x}{the list to examine.}
}
\value{
\code{\link[=invisible]{invisible(TRUE)}}
}
\description{
Returns \code{TRUE} if each element in \code{x} has length 1. Otherwise it will produce
an error with \code{\link{stop}}.
}
\seealso{
\code{\link[=stop]{stop()}}
}
