% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pastec.R
\name{pastec}
\alias{pastec}
\alias{pastec0}
\title{Concatenate strings sequentially}
\usage{
pastec(..., sep = " ", collapse = " ")

pastec0(...)
}
\arguments{
\item{...}{one or more \code{R} objects, to be converted to character vectors.}

\item{sep}{a character string to separate the terms. Not \code{\link{NA_character_}}.}

\item{collapse}{an optional character string to separate the results. Not
\code{\link{NA_character_}}.}
}
\description{
\code{pastec} and \code{pastec0} recursively combine their variable arguments with
\code{\link{c}} before passing it to \code{\link{paste}} or \code{\link{paste0}}, respectively.
}
\details{
\code{pastec} collapses and separates objects with a space by default but can be
customized, while \code{pastec0} always collapses with the empty string.
}
\seealso{
\code{\link[=paste]{paste()}} \code{\link[=paste0]{paste0()}}
}
