% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_repo.R
\name{set_repo}
\alias{set_repo}
\title{Set CRAN repository}
\usage{
set_repo(repo_url)
}
\arguments{
\item{repo_url}{The URL of the CRAN repository to download packages from}
}
\value{
The previously repository URL, invisibly.
}
\description{
\code{set_repo} sets the URL of the repository to download new packages from.
}
\details{
This sets \code{options(repos)}, overwriting the value. The previous value is
returned, which can be helpful if you want to set the repository only
temporarily.
}
